Add ShapeFactory extensions (#16715)
---------
